Refactor many of the things
