+++ src/pybullet_helpers/__init__.py
@@ -1,0 +1,6 @@
+"""This implementation is heavily based on the pybullet-planning repository by
+Caelan Garrett (https://github.com/caelan/pybullet-planning/).
+
+In addition, the structure is loosely based off the pb_robot repository
+by Rachel Holladay (https://github.com/rachelholladay/pb_robot).
+"""
