+++ src/pybullet_helpers/ikfast/__init__.py
@@ -1,5 +1,3 @@
-"""This implementation is heavily based on the pybullet-planning repository by
-Caelan Garrett (https://github.com/caelan/pybullet-planning/)."""
 from typing import List, NamedTuple
