+++ src/pybullet_helpers/robots/__init__.py
@@ -2,9 +2,9 @@ import importlib
 import pkgutil
 from typing import TYPE_CHECKING, Dict
 
+from predicators.src.pybullet_helpers.geometry import Pose
 from predicators.src.pybullet_helpers.robots.single_arm import \
     SingleArmPyBulletRobot
-from predicators.src.pybullet_helpers.utils import Pose
 from predicators.src.structs import Pose3D, Quaternion
 from predicators.src.utils import get_all_concrete_subclasses
 
@@ -18,8 +18,8 @@ if not TYPE_CHECKING:
 
 # Note: these are static poses which suffice for the current environments.
 _ROBOT_TO_BASE_POSE: Dict[str, Pose] = {
-    "fetch": Pose((0.75, 0.7441, 0.0)),
-    "panda": Pose((0.8, 0.7441, 0.25)),
+    "fetch": Pose(position=(0.75, 0.7441, 0.0)),
+    "panda": Pose(position=(0.8, 0.7441, 0.25)),
 }
 
 
@@ -40,6 +40,6 @@ def create_single_arm_pybullet_robot(
                         base_pose=base_pose)
             break
     else:
-        raise NotImplementedError(f"Unrecognized robot name: {robot_name}.")
+        raise ValueError(f"Unrecognized robot name: {robot_name}.")
 
     return robot
